First release of new ECAL condition object EcalWeightRecAlgoWeights -------------------------------------------------------------------
this is a container class that will be made persistent and available to
reco clients through the unified source mechanism

the internal objects are of the type vector< vector<EcalWeight> >
where EcalWeight is simply a wrapper for double!
This is because of classes beuween object in the dictionary from
framework and this package.

once the problem is solved EcalWeight will become simply a typedef for double!