Added first version of the analytic fit as implemented by Alessandro Palma
Buildfile has now dependency on root and has to use explicitly the Minuit library

the dummy file has dummy instances of the algorithms in order to handle dependencies of
packages with only templated classes. THis is due to a bug in scram and will be cleaned up
once the bug is fixed.