adding new cond db objects for the ecal:
EcalADCToGeVConstant: global energy scale factor to convert ADC -> GeV

EcalMGPAGainRatio: gain ratios 12/6 and 6/1 for individual channels

EcalGainRatios: set of EcalMGPAGainRatio for all ecal xtals

EcalXtalGroupId: simple id class to group togethers subsets of similar xtals,
for example those with similar timing

EcalWeightXtalGroups: map of which ctals are in which group. used to store only
a small number  of sets of weights.

EcalIntercalibConstants: map of one constant / channel to be used
for intercalibrtation

EcalWeightSet: replaces the old EcalWeightRecAlgoWeights as the basi set of all numbers
defined as weights.

EcalTBWeights: map of EcalTBWeights for each pair of (EcalXtalGroupId, TDC info)