Method driftDirection() has a bug in the x and y component. As a result, lorentzShiftX() and lorentzShiftY() come out with a wrong sign.  (This bug existed in Orca, and has already been fixed in the Digitizer, but not here.) However, its impact is small since chargeWidth() adds the geometrical correction to the incorrectly signed lorentz drift and then takes an absolute value of the sum -- which largely masks the bug.
In order to study this effect, I (Petar) am leaving the incorrect code
in, and adding another function computeLorentzShifts().  This one does
both in one go (thus saving a couple of calls) and also leaves the
values of the lorentz drifts in X and Y in cm too -- since CPE Generic
needs these values in the local units anyway.