unit tests of statistics at end of ordinary file.
unit test of setting limit-default for a severity (INFO) to filter
all the info messages other than the FwkReport messages from the
cerr stream in the default configuration.

correction of a typo in MessageLoggerScribe.cc that would have prevented
that filtering from working (expressly for cerr, and no other destination!)

change in MessageLogger.cfi to take advantage of that ability.