As also posted on the pixel software hypernews fora.
There is now a working version of the CMSSW module IORawData/SiPixelInputSources that can run on files in castor! For anyone interested in this particular feature: please upgrade IORawData/SiPixelInputSources to tag V00-00-12. This version will hopefully also be queued for the CMSS 17X series (however, it can actually be checked out and used in earlier releases like the 16X tree).

However, there are some small caveats (that considering the typical use I've observed up to now should not stop us much)

1. The module can only read one (1) single file from castor, it is not possible to give a list of files
2. For CMSSW-standards, the module is relatively slow. This has to do with the fact that each separate event has to be streamed, and the rfio protocol is very much tuned at reading large buffers. It would be better to read a large buffer (32K even more) and read then events from there, but unfortunately rfio does not support directly buffered-I/O as it is also used for random-reads (such as root). If someone wants to volunteer, it would be possible to write the buffering code... but it won't be me.
3. Caching very large files takes long (25 Mbytes/sec is normal, so we can all calculate that our 'typical' S-curve files (>10 Gbyte) should take several minutes to cache). However, this is probably still preferable over copying everything somewhere else from castor. The grid/castor community is aware of this but I have not heard any suggestions on how to improve this (I'll keep you posted)