Allowed delayed setting of the MagneticField: 1) default in PixelCPEBase constructor for the magfield is now 0 2) there is a setMagneticField() method 3) magnetic field is now a mutable const to allow (2)
This now allows CPE's to be built first (e.g. in the constructor)
and the magnetic field obtained from the EventSetup later on.
It also decouples CPE from Magnetic Field which should simplify
obtaining both independently from the EventSetup later on.

Also, changing the computation of cotalpha_ and cotbeta_ from
1/tan(angle), which is unsafe at direct incidence, to
tan(pi/2-angle) which is fine.