Eliminated the static object that was causing stand-alone programs which linked MessageLogger but never established its presence to report a problem (involving thread usage) after main exited. Tested via trivial_main.cpp.
Forbid copying objects like LogInfo, since the behavior when that is
done would be confusing at best.

Move the static free function onlyLowestDirectory into a class function
and out of the header, so that the observed many-copies-of-this-function
performance worry is cured.

Added a .get() method to value_ptr, which was an inadvertant omission.