Include target variable in the ProcMatrix correlation plot
Implement variable ranking as a side-feature of ProcMatrix:
Use ranking="true" in ProcMatrix configuration.  No output variables,
but ranking is done.  Results available in ViewMonitoring.C
(or as a histogram in the monitoring ROOT file).  Additional gained
"correlation to target" in percent for each variable is listed.
This information is gained from an iterative procedure.  The worst
variable is iteratively removed, starting from all included variables
and the "loss in correlation to the target" quantified in each step.

Input variables should be linearized with respect
to the target, i.e. passed through ProcLikelihood individiually like
this:

        <processor id="rank" name="ProcLikelihood">
                <input>
                        <var source="norm" name="x"/>
                        <var source="norm" name="y"/>
                </input>
                <config>
                        <general individual="true"/>
                        <sigbkg/>
                        <sigbkg/>
                </config>
                <output>
                        <var name="x"/>
                        <var name="y"/>
                </output>
        </processor>
        <processor id="ranking" name="ProcMatrix">
                <input>
                        <var source="rank" name="x"/>
                        <var source="rank" name="y"/>
                </input>
                <config>
                        <fill ranking="true"/>
                </config>
                <output/>
        </processor>