Gracefully handle bad (e.g. truncated) configuration files.
In case the bad configuration was specified on command-line, we assume that the
user knew what he was doing and exit with an error message.

In case the bad configuration was selected via the dialog, we pop up a modal
dialog, reset everything and expect the user to pick a new configuration.