Fix placement of windows -- hopefully for all window managers.
Add function
   void measureWMOffsets();
and data-members
   int m_WMOffsetX, m_WMOffsetY, m_WMDecorH;
to store window-manager related information.
This is acquired once at construction time by opening a window
and storing its displacement into the data-members.

Make void setWindowInfoFrom() a member-function as it needs to access the
offsets (was static, local to compilation unit before).

Make sure that y-position of the window is larger than the decor height.