Fix Error Calculation
Asymmetric error approach was taken in the previous version of the code with
sigma up(down) calculated as the mean value of sum of the squared of up (down)
deviations from the central value.

New code runs given number of pseudoexperiments, plot calculated values of the
efficiencies, fit with gaussian and use sigma of the distribution as the
symmetric error.