Rename dqmFormat parameter passed to the merging step into newDQMIO. Its meaning also changed: if the parameter is present and True the new DQM IO format will be properly configured for the merging. If the parameter is absent or False, the old format will be used.
Updated all scenarios to be able to dynamically select between new and old
DQM IO format. The output module is automatically configured by configBuilder
everytime the DQMROOT dataTier is specified.

The ENDJOB sequence has been put back in place for all scenarios, since
configBulider will take care of removing it in case the DQMROOT dataTier is
selected.

The harvesting sequence will dynamically use the new/old IO format depending
on the presence/value of the parameter newDQMIO that must be passed to the
harvesting method.