- Implemented all public lumi plots for 2013 proton-lead. - Added a config file for 2013 proton-lead plots. - Fixed a small bug in determination of the vertical axis range for log scale plots. - Fixed a bug in the binning into weeks.
NOTE: Both the code and the config file still contain a workaround for the fact
that due to a problem somewhere the accelerator mode is not stored correctly
in the RunInfo database (from where lumiCalc gets this information). Using the
'PAPHYS' amodetag results in no data being returned at all.